Added try block to handle exceptions in SetVersionNumbersInProjCsproj_FilesAsync
A `try` block has been added to the `SetVersionNumbersInProjCsproj_FilesAsync` method in the `SetGlobalVersion` namespace in the `MyToolWindowControl.xaml.cs` file. This change is designed to handle potential errors or exceptions that may occur when checking if a file exists at the specified path. The `catch` block will handle any exceptions thrown.